% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr2lang.R
\name{chr2lang}
\alias{chr2lang}
\title{Vectorized \code{str2lang}}
\usage{
chr2lang(x, .named = TRUE)
}
\arguments{
\item{x}{a character vector.}

\item{.named}{a logical.}
}
\description{
Vectorized \code{str2lang}
}
