% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enlang.R
\name{enlang}
\alias{enlang}
\title{Convert an object(s) to an empty \code{call_simple}}
\usage{
enlang(x)
}
\arguments{
\item{x}{an R object.}

\item{...}{R objects.}

\item{.named}{If TRUE, unnamed inputs are automatically named with \code{as_label()}. This is equivalent to applying \code{exprs_auto_name()} on the result. If FALSE, unnamed elements are left as is and, if fully unnamed, the list is given minimal names (a vector of ""). If NULL, fully unnamed results are left with NULL names.}

\item{.ignore_empty}{Whether to ignore empty arguments. Can be one of "trailing", "none", "all". If "trailing", only the last argument is ignored if it is empty. Named arguments are not considered empty.}

\item{.unquote_names}{Whether to treat ⁠:=⁠ as =. Unlike =, the ⁠:=⁠ syntax supports names injection.}

\item{.ignore_null}{Whether to ignore unnamed null arguments. Can be "none" or "all".}

\item{.homonyms}{How to treat arguments with the same name. The default, "keep", preserves these arguments. Set .homonyms to "first" to only keep the first occurrences, to "last" to keep the last occurrences, and to "error" to raise an informative error and indicate what arguments have duplicated names.}

\item{.check_assign}{Whether to check for ⁠<-⁠ calls. When TRUE a warning recommends users to use = if they meant to match a function parameter or wrap the ⁠<-⁠ call in curly braces otherwise. This ensures assignments are explicit.}
}
\value{
a single or a list of \code{call_simple}.
}
\description{
Convert an object(s) to an empty \code{call_simple}
}
\examples{
enlang(mean)
}
