% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_compose.R
\name{fn_compose}
\alias{fn_compose}
\title{Compose and evaluate functions}
\usage{
fn_compose(.x = NULL, ..., .dir = c("forward", "backward"))
}
\arguments{
\item{.x}{a vector.}

\item{...}{functions or formulas (lambda functions).}

\item{.dir}{composition direction. Defaults to \code{forward} (unlike \code{purrr::compose}). Alternative is \code{backward}.}
}
\value{
a list.
}
\description{
Compose and evaluate functions
}
