% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ns_call.R
\name{is_ns_call}
\alias{is_ns_call}
\alias{is_ns_call_export}
\alias{is_ns_call_private}
\title{Title}
\usage{
is_ns_call(x, ns_type = c("all", "export", "private"))

is_ns_call_export(x)

is_ns_call_private(x)
}
\arguments{
\item{x}{a \code{call_simple}.}

\item{ns_type}{a string.}
}
\value{
a logical.
}
\description{
Title
}
\examples{
x0 <- quote(mean())
x1 <- quote(base::mean())
x2 <- quote(base:::mean())

is_ns_call(x0)
is_ns_call(x1)
is_ns_call(x2)

is_ns_call(x1, "export")
is_ns_call(x1, "private")

is_ns_call(x2, "export")
is_ns_call(x2, "private")

}
