% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_numchar.R
\name{is_numchar}
\alias{is_numchar}
\title{Is an object a numeric vetor?}
\usage{
is_numchar(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
a logical.
}
\description{
Is an object a numeric vetor?
}
