% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_symbol2.R
\name{is_symbol2}
\alias{is_symbol2}
\title{Is object a symbol?}
\usage{
is_symbol2(x, .ns_call = TRUE)
}
\arguments{
\item{x}{An object to test.}

\item{.ns_call}{a logical.}
}
\description{
Is object a symbol?
}
