% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang2fun.R
\name{lang2fun}
\alias{lang2fun}
\title{from expression to function}
\usage{
lang2fun(x, env = globalenv())
}
\arguments{
\item{x}{a symbol, a call, a string, or a function.}

\item{env}{environment. Defaults to global environment.}
}
\description{
from expression to function
}
