% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objs_as_call.R
\name{objs_as_call}
\alias{objs_as_call}
\title{Convert strings, symbols, and calls into \code{call_simple}}
\usage{
objs_as_call(..., enexpr_input = TRUE, ns_sym = TRUE, str_as_expr = TRUE)
}
\arguments{
\item{...}{a string, symbol, or call.}

\item{enexpr_input}{logical. If \code{TRUE}, \code{rlang::enexpr} is applied to input.Defaults to \code{TRUE}.}

\item{ns_sym}{logical. If \code{FALSE}, only symbols are converted into \code{call_simple}. If \code{TRUE}, both \code{symbol} and \code{namespaced_symbol} (\verb{namespace::function}). Defaults to \code{TRUE}.}

\item{str_as_expr}{logical. if \code{TRUE}, strings are quoted. Not if \code{FALSE}. Defaults to \code{TRUE}.}
}
\value{
a \code{call_simple}.
}
\description{
Convert strings, symbols, and calls into \code{call_simple}
}
\examples{
objs_as_call(mean)
objs_as_call("mean")
objs_as_call(mean())
objs_as_call("mean()")
objs_as_call(base::mean)
objs_as_call("base::mean")
}
