% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste2.R
\name{paste2}
\alias{paste2}
\title{Variations of \code{paste}}
\usage{
paste2(..., .before = NULL, sep = "", collapse = NULL, recycle0 = FALSE)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}

\item{.before}{a character vecror.}

\item{sep}{a character string to separate the terms. Not NA_character_.}

\item{collapse}{an optional character string to separate the results. Not NA_character_. When collapse is a string, the result is always a string (character of length 1).}

\item{recycle0}{logical indicating if zero-length character arguments should result in the zero-length chara}
}
\value{
A character vector of the concatenated values.
}
\description{
Variations of \code{paste}
}
