% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf2.R
\name{sprintf2}
\alias{sprintf2}
\title{Variation of \code{sprintf}}
\usage{
sprintf2(..., fmt, .before = NULL)
}
\arguments{
\item{...}{values to be passed into fmt.}

\item{fmt}{a character vector of format strings, each of up to 8192 bytes.}

\item{.before}{a character vector.}
}
\description{
Variation of \code{sprintf}
}
